\name{IBL015747.db}
\alias{IBL015747.db}
\alias{IBL015747}
\title{Bioconductor annotation data package}
\description{
  
  Welcome to the IBL015747.db annotation Package.  The purpose of
  this package is to provide detailed information about the
  IBL015747 platform. This package is updated biannually.
  
  You can learn what objects this package supports with the following command:

  \code{ls("package:IBL015747.db")} 
  
  Each of these objects has their own manual page detailing where
  relevant data was obtained along with some examples of how to use it.
}
\examples{
  ls("package:IBL015747.db")
}
\keyword{datasets}

